Separated MPI and calc_wiggled, now running
