+++ Makefile
@@ -5,5 +5,5 @@ integrate:
 	python3 -m unittest discover -s tests.integration -v
 
 parallelwiggle:
-	mpiexec -n 5 python3 -m unittest tests.unit.test_calc_wiggled -v
+	mpiexec -n 5 python3 -m superbol.mpi_shell -v
